+++ pages/palvelumme.js
@@ -1,10 +1,7 @@
 import React, { useEffect } from 'react';
 import Head from 'next/head';
-import { Navbar } from '../src/komponentit/Navbar/Navbar'
-import { PalvelutLuettelo } from '../src/komponentit/PalvelutLuettelo/PalvelutLuettelo'
-import { TarjousForm } from '../src/komponentit/TarjousForm/TarjousForm'
-import { Footer } from '../src/komponentit/Footer/Footer'
-import Cookies from '../src/komponentit/Cookies';
+import { PalvelutLuettelo } from '../src/komponentit/PalvelutLuettelo'
+import { TarjousForm } from '../src/komponentit/TarjousForm'
 
 export const Palvelumme = () => {
