+++ pages/referenssit.js
@@ -1,9 +1,6 @@
 import React, {useEffect} from 'react';
 import Head from 'next/head';
-import { Navbar } from '../src/komponentit/Navbar/Navbar'
-import { ReferenssiKohteet } from '../src/komponentit/ReferenssiKohteet/ReferenssiKohteet'
-import { Footer } from '../src/komponentit/Footer/Footer'
-import Cookies from '../src/komponentit/Cookies';
+import { ReferenssiKohteet } from '../src/komponentit/ReferenssiKohteet'
 
 export const Referenssit = () => {
