+++ pages/tietosuojaseloste.js
@@ -1,9 +1,6 @@
 import React, { useEffect } from 'react'
 import Head from 'next/head'
-import { Navbar } from '../src/komponentit/Navbar/Navbar'
-import { TietosuojaselosteContent } from '../src/komponentit/TietosuojaselosteContent/TietosuojaselosteContent'
-import { Footer } from '../src/komponentit/Footer/Footer'
-import Cookies from '../src/komponentit/Cookies'
+import { TietosuojaselosteContent } from '../src/komponentit/TietosuojaselosteContent'
 
 export const Tietosuojaseloste = () => {
