--- src/komponentit/TietosuojaselosteContent/TietosuojaselosteContent.js
@@ -1,71 +1,0 @@
-import React from 'react'
-
-export const TietosuojaselosteContent = () => {
-  return (
-    <div className='tietosuojaseloste-container'>
-        <h1>Tietosuojaseloste</h1>
-        <div className='tietosuojaseloste-content'>
-
-            <h2>1. Rekisteri</h2><br />
-            <p>Verkkosivujen yhteydenottolomakkeen sekä asiakkaalta 
-                saatujen asennusprosessin kannalta oleellisten henkilötietojen rekisteri.</p><br />
-
-            <h2>2. Rekisterinpitäjä</h2><br />
-            <p>
-                Vihergia Oy<br /><br />
-            Y-tunnus: 3333004-9<br /><br />
-            Osoite: Sippolankuja 13<br /><br />
-            Postinumero: 04430<br /><br />
-            Postitoimipaikka: Järvenpää<br /><br />
-            Puhelinnumero: 044 088 3886<br /><br />
-            Sähköpostiosoite: info@vihergia.fi<br /><br />
-            </p>
-            <h2>3. Yhteyshenkilö rekisteriä koskevissa asioissa</h2><br />
-            <p>
-                Nimi: Toni Pystö<br /><br />
-                Y-tunnus: 3333004-9<br /><br />
-                Osoite: Sippolankuja 13<br /><br />
-                Postinumero 04430<br /><br />
-                Postitoimipaikka: Järvenpää<br /><br />
-                Puhelinnumero: 044 088 3886<br /><br />
-                Sähköpostiosoite: toni.pysto@vihergia.fi<br /><br />
-            </p>
-            <br />
-
-            <h2>4. Henkilötietojen käsittelyn tarkoitus</h2><br />
-
-            <p>Yhteydenottolomakkeella ja asiakkaalta muuta kautta tulevista tiedoista kerättävän rekisterin tarkoituksena on edistää palvelun tuottamista asiakkaalle. <br />Rekisterin tietoja käytetään tarjousten ja sopimusten lähettämiseen asiakkaalle, asennuksen järjestämiseen, laskutukseen, mahdollisiin takuukorjauksiin sekä asiakkaan tietojen lähettämiseen sähköverkkoyhtiölle aurinkosähköjärjestelmän ilmoituksia varten.</p><br />
-
-            <h2>5. Rekisterin tietosisältö</h2><br />
-            <p>Rekisteriin tallennetaan asiakkaalta verkkosivujen yhteydenottolomakkeella sekä muuta kautta tullut asioiden hoitamisen kannalta oleellinen tieto. Näitä ovat henkilön tai yrityksen nimi, osoite, puhelinnumero, sähköpostiosoite.</p><br />
-            <h2>6. Tietojen säilytysaika</h2><br />
-            <p>Henkilötietoja säilytetään niin kauan kuin on tarve asiakkaan kanssa tehdyn sopimuksen täyteenpanemiseksi. Mahdollisten takuukorjausten kannalta oleellisia tietoja säilytetään asennuksen takuuajan umpeutumiseen saakka.</p><br />
-
-            <h2>7. Säännönmukaiset tietolähteet</h2><br />
-
-            <p>Rekisterissä olevia tietoja kerätään verkkosivujen yhteydenottolomakkeella sekä henkilöltä itseltään. Tietoja voidaan kerätä myös viranomaisten pitämistä rekistereistä (esim. YTJ) lain sallimissa rajoissa.</p><br />
-
-            <h2>8. Tietojen säännönmukaiset luovutukset</h2><br />
-
-            <p>Tietoja ei säännönmukaisesti luovuteta organisaation ulkopuolelle. Tietoja kerrotaan tai luovutetaan vain lakiin perustuvan ilmoitusvelvollisuuden johdosta.</p><br />
-
-            <h2>9. Tietojen siirto EU:n tai EFTA:n ulkopuolelle</h2><br />
-
-            <p>Henkilötietoja ei siirretä EU:n tai EFTA:n ulkopuolelle.</p><br />
-
-            <h2>10. Rekisterin suojaus</h2><br />
-
-            <p>Sähköinen aineisto on suojattu henkilökohtaisilla käyttäjätunnuksilla ja salasanoilla. Tietojen käyttöoikeus on vain niillä Vihergia Oy:n palveluksessa olevilla henkilöillä, jotka tarvitsevat tietoja työn suorittamiseksi. Sähköisessä muodossa olevaa tietoa tulostetaan vain tarvittaessa, niitä säilytetään lukitussa tilassa ja ne tuhotaan välittömästi käytön jälkeen.</p><br />
-
-            <h2>11. Rekisteröidyn oikeudet</h2><br />
-            <p>
-                Rekisteröidyllä on oikeus tarkastaa hänestä rekisteriin merkityt tiedot, vaatia virheellisten tietojen muutosta, sekä turvautua laissa turvattuihin oikeuksiinsa. Henkilötietojen käsittely perustuu rekisteröidyn antamaan suostumukseen ja rekisteröidyllä on oikeus milloin tahansa perua antamansa suostumus ja pyytää tietojensa poistamista.<br />
-
-                Mikäli rekisteröity haluaa tarkastaa, muuttaa tai poistaa rekisterissä olevia tietojaan, tulee hänen lähettää kirjallinen pyyntö rekisterinpitäjälle omakätisesti allekirjoitettuna tai muulla tavalla varmennettuna.
-            </p>
-        </div>
-    </div>
-  )
-}
-
-export default TietosuojaselosteContent
