+++ src/komponentit/layout.js
@@ -1,5 +1,5 @@
-import Navbar from "./Navbar/Navbar"
-import Footer from "./Footer/Footer"
+import Navbar from "./Navbar"
+import Footer from "./Footer"
 import Cookies from "./Cookies"
 
 export default function Layout({ children }) {
